requirements.txt
click
python-can
simplejson
